Moved example plugin EP's kernel stuff to different file
